+++ preparation/git_clone.py
@@ -1,6 +1,6 @@
 from git.repo import Repo
 
-from conf import LLVM_GIT_URL, LLVM_PROJ_NAME, SRC_ROOT
+from conf.conf import LLVM_GIT_URL, LLVM_PROJ_NAME, SRC_ROOT
 from utils.record_step import clang_cloned, compiler_rt_elf_cloned, \
     compiler_rt_musl_cloned, is_clang_cloned, libcxx_cloned, libcxxabi_cloned, \
     libunwind_cloned
